requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==22.0.0
SQLAlchemy==2.0.36
aiomysql==0.2.0
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.6.0
passlib[argon2]==1.7.4
pyjwt==2.9.0
python-multipart==0.0.9
httpx==0.27.2
orjson==3.10.7
prometheus-client==0.20.0
email-validator==2.2.0

